+++ sidechain_cli/utils/config_file/__init__.py
@@ -1,0 +1,16 @@
+"""Util methods for the sidechain CLI."""
+
+from sidechain_cli.utils.config_file.bridge_config import BridgeConfig
+from sidechain_cli.utils.config_file.chain_config import ChainConfig
+from sidechain_cli.utils.config_file.config_file import ConfigFile, get_config_folder
+from sidechain_cli.utils.config_file.server_config import ServerConfig
+from sidechain_cli.utils.config_file.witness_config import WitnessConfig
+
+__all__ = [
+    "BridgeConfig",
+    "ChainConfig",
+    "WitnessConfig",
+    "ServerConfig",
+    "ConfigFile",
+    "get_config_folder",
+]
